values size is long int

+++ mpi/mpi.cpp
@@ -10,7 +10,7 @@ using namespace std;
 int main(int argc, char *argv[]) {
 	int proc_number, world_rank;
 
-	long long unsigned int partial_sum, global_sum;
+	double partial_sum, global_sum;
 	partial_sum = global_sum = 0;
 
 	//Initialize the MPI environment
@@ -30,9 +30,9 @@ int main(int argc, char *argv[]) {
 	//Reading size of  matrix from file
 	int n;
 	FILE *file = fopen("matrix","rb");
-	fread(&n,sizeof(int),1,file);
+	fread(&n,sizeof(double),1,file);
 
-	int values_size = n*n/(proc_number-1); //this 1 is the master node
+	long int values_size = n*n/(proc_number-1); //this 1 is the master node
 
 	//variables for the time
 	struct timeval stop, start;
@@ -44,16 +44,16 @@ int main(int argc, char *argv[]) {
 
 		//allocating memory for the matrix
 		//reading the matrix from the file
-		int *buffer = (int*)malloc(sizeof(int)*(n*n+1));
-		fread(buffer,sizeof(int),n*n,file);
+		double *buffer = (double*)malloc(sizeof(double)*(n*n+1));
+		fread(buffer,sizeof(double),n*n,file);
 	
 		//Starting the time after reading the file
 		gettimeofday(&start, NULL);
 		start_seconds = start.tv_sec + start.tv_usec*1e-6;
 
 		//Sending the diagonal of the matrix
-		int *superior_diagonal = new int[n/2];
-		int *inferior_diagonal = new int[n/2];
+		double *superior_diagonal = new double[n/2];
+		double *inferior_diagonal = new double[n/2];
 		int i = 0;
 		for(i = 0; i < n; i++){
 			if (i < n/2)
@@ -63,13 +63,13 @@ int main(int argc, char *argv[]) {
 		}
 
 		if(proc_number == 3){
-			MPI_Send(superior_diagonal, n/2, MPI_INT, 1, 0, MPI_COMM_WORLD);
-			MPI_Send(inferior_diagonal, n/2, MPI_INT, 2, 0, MPI_COMM_WORLD);
+			MPI_Send(superior_diagonal, n/2, MPI_DOUBLE, 1, 0, MPI_COMM_WORLD);
+			MPI_Send(inferior_diagonal, n/2, MPI_DOUBLE, 2, 0, MPI_COMM_WORLD);
 		} else if (proc_number == 5){
-			MPI_Send(superior_diagonal, n/2, MPI_INT, 1, 0, MPI_COMM_WORLD);
-			MPI_Send(superior_diagonal, n/2, MPI_INT, 2, 0, MPI_COMM_WORLD);
-			MPI_Send(inferior_diagonal, n/2, MPI_INT, 3, 0, MPI_COMM_WORLD);
-			MPI_Send(inferior_diagonal, n/2, MPI_INT, 4, 0, MPI_COMM_WORLD);
+			MPI_Send(superior_diagonal, n/2, MPI_DOUBLE, 1, 0, MPI_COMM_WORLD);
+			MPI_Send(superior_diagonal, n/2, MPI_DOUBLE, 2, 0, MPI_COMM_WORLD);
+			MPI_Send(inferior_diagonal, n/2, MPI_DOUBLE, 3, 0, MPI_COMM_WORLD);
+			MPI_Send(inferior_diagonal, n/2, MPI_DOUBLE, 4, 0, MPI_COMM_WORLD);
 		} else {
 			cout << "Número inválido de processadores." << endl;
 			return 1;
@@ -77,16 +77,16 @@ int main(int argc, char *argv[]) {
 
 		//Splitting and sending each part of the matrix
 		if(proc_number == 3){
-			int *values_to_send1 = buffer;
-			int *values_to_send2 = buffer + values_size;
+			double *values_to_send1 = buffer;
+			double *values_to_send2 = buffer + values_size;
 
-			MPI_Send(values_to_send1, values_size, MPI_INT, 1, 0, MPI_COMM_WORLD);
-			MPI_Send(values_to_send2, values_size, MPI_INT, 2, 0, MPI_COMM_WORLD);
+			MPI_Send(values_to_send1, values_size, MPI_DOUBLE, 1, 0, MPI_COMM_WORLD);
+			MPI_Send(values_to_send2, values_size, MPI_DOUBLE, 2, 0, MPI_COMM_WORLD);
 		} else if (proc_number == 5){
-			int *values_to_send1 = new int[values_size];
-			int *values_to_send2 = new int[values_size];
-			int *values_to_send3 = new int[values_size];
-			int *values_to_send4 = new int[values_size];
+			double *values_to_send1 = new double[values_size];
+			double *values_to_send2 = new double[values_size];
+			double *values_to_send3 = new double[values_size];
+			double *values_to_send4 = new double[values_size];
 
 			int i, j;
 			for (i = 0; i < n/2; i++){
@@ -98,20 +98,20 @@ int main(int argc, char *argv[]) {
 				}
 			}
 
-			MPI_Send(values_to_send1, values_size, MPI_INT, 1, 0, MPI_COMM_WORLD);
-			MPI_Send(values_to_send2, values_size, MPI_INT, 2, 0, MPI_COMM_WORLD);
-			MPI_Send(values_to_send3, values_size, MPI_INT, 3, 0, MPI_COMM_WORLD);
-			MPI_Send(values_to_send4, values_size, MPI_INT, 4, 0, MPI_COMM_WORLD);
+			MPI_Send(values_to_send1, values_size, MPI_DOUBLE, 1, 0, MPI_COMM_WORLD);
+			MPI_Send(values_to_send2, values_size, MPI_DOUBLE, 2, 0, MPI_COMM_WORLD);
+			MPI_Send(values_to_send3, values_size, MPI_DOUBLE, 3, 0, MPI_COMM_WORLD);
+			MPI_Send(values_to_send4, values_size, MPI_DOUBLE, 4, 0, MPI_COMM_WORLD);
 		}
 	} else { //These are the others nodes
-		int received_diagonal[n/2];
-		MPI_Recv(received_diagonal, n/2, MPI_INT, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
+		double received_diagonal[n/2];
+		MPI_Recv(received_diagonal, n/2, MPI_DOUBLE, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
 
-		int received_values[values_size];
-		MPI_Recv(received_values, values_size, MPI_INT, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
+		double received_values[values_size];
+		MPI_Recv(received_values, values_size, MPI_DOUBLE, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
 
 		//Calculating the partial_sum
-		int new_values[values_size];
+		double new_values[values_size];
 		int columns_number = n/((proc_number-1)/2);
 		int i,j;
 		for(i = 0; i < n/2; i++)
@@ -122,7 +122,7 @@ int main(int argc, char *argv[]) {
 	}
 
 	//Using MPI_Reduce to sum the partial_sum of each node
-	MPI_Reduce(&partial_sum, &global_sum, 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+	MPI_Reduce(&partial_sum, &global_sum, 1, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
 	gettimeofday(&stop, NULL);
 	stop_seconds = stop.tv_sec + stop.tv_usec*1e-6;
